+++ klein-consensus/klein-consensus-facade/src/main/java/com/ofcoder/klein/consensus/facade/ConsensusEngine.java
@@ -1,28 +1,22 @@
 package com.ofcoder.klein.consensus.facade;
 
-import com.ofcoder.klein.consensus.facade.config.ConsensusProp;
-import com.ofcoder.klein.spi.ExtensionLoader;
-import com.ofcoder.klein.spi.SPI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import com.ofcoder.klein.consensus.facade.config.ConsensusProp;
+import com.ofcoder.klein.spi.ExtensionLoader;
 
 /**
  * @author far.liu
  */
-@SPI
 public class ConsensusEngine {
     private static final Logger LOG = LoggerFactory.getLogger(ConsensusEngine.class);
     private static Consensus consensus;
 
     public static void startup(String algorithm, ConsensusProp prop) {
         LOG.info("start consensus engine");
-
         consensus = ExtensionLoader.getExtensionLoader(Consensus.class).getJoin(algorithm);
         consensus.init(prop);
-    }
-
-    public static void loadSM(SM sm) {
-        consensus.loadSM(sm);
     }
 
     public static void shutdown() {
