+++ klein-consensus/klein-consensus-paxos/src/main/java/com/ofcoder/klein/consensus/paxos/PaxosConsensus.java
@@ -27,7 +27,7 @@ import org.slf4j.LoggerFactory;
 import com.ofcoder.klein.consensus.facade.Consensus;
 import com.ofcoder.klein.consensus.facade.MemberManager;
 import com.ofcoder.klein.consensus.facade.Result;
-import com.ofcoder.klein.consensus.facade.SM;
+import com.ofcoder.klein.consensus.facade.sm.SM;
 import com.ofcoder.klein.consensus.facade.config.ConsensusProp;
 import com.ofcoder.klein.consensus.facade.exception.ConsensusException;
 import com.ofcoder.klein.consensus.paxos.core.Acceptor;
