+++ klein-core/src/main/java/com/ofcoder/klein/core/cache/CacheSM.java
@@ -23,7 +23,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.collect.ImmutableMap;
-import com.ofcoder.klein.consensus.facade.AbstractSM;
+import com.ofcoder.klein.consensus.facade.sm.AbstractSM;
 
 /**
  * @author 释慧利
