+++ readme.md
@@ -12,22 +12,23 @@ paxos, grpc, h2
 
 ## 进度
 ### paxos
-1. [x] 写请求、乱序协商，顺序确认
-2. [x] 读请求，使用协商log完成
-3. [x] 批量协商
-4. [x] 优化prepare阶段
-5. [x] 快照
-6. [ ] 增加Master：成员变更、优化读请求
-7. [ ] 成员自动发现
-8. [ ] 数据对齐：成员上线、落后成员对齐
-9. [ ] NWR
-10. [ ] confirm优化读请求
-11. [ ] 不存在干扰key，无需执行一轮Prepare
+- [x] 写请求、乱序协商，顺序确认
+- [x] 读请求，使用协商log完成
+- [x] 批量协商
+- [x] 优化prepare阶段
+- [x] 快照
+- [ ] 拆分Group，proposer等角色无须隔离，只需隔离instance
+- [ ] 增加Master：成员变更、优化读请求
+- [ ] 成员自动发现(调研)
+- [ ] 数据对齐：成员上线、落后成员对齐
+- [ ] NWR
+- [ ] confirm优化读请求
+- [ ] 不存在干扰key，无需执行一轮Prepare
 
 ### 缓存
-1. [x] 读、写、等基础功能
-2. [ ] 配合持久化实现LRU
-3. [ ] TTL自动过期
+- [x] 读、写、等基础功能
+- [ ] 配合持久化实现LRU
+- [ ] TTL自动过期
 
 # 章解
 - 批量协商（队列），减少Prepare阶段，减少RPC交互
